Added git subcommand `ahead-behind` with wrapper
Usually it's just useful to view the commits ahead/behind the remote
branch until you are submitting a PR, but constantly being on a
different branch per device makes it useful to compare these branches
to master locally. If there's a way to do this simply through a
subcommand already then I am not aware.

Not sure if a git wrapper is too dangerous, might want to just make
this an aliased function.